Unify test suite
Making all tests under doublex.test run on Python 3 only.
