+++ doublex/test/unit_tests.py
@@ -1716,6 +1716,19 @@ class TracerTests(TestCase):
         assert_that(self.out.getvalue(),
                     is_("ObjCollaborator.prop set to 2"))
 
+
+# new on 1.8.3
+# issue: https://bitbucket.org/DavidVilla/python-doublex/issues/25/support-from-python-35-type-hints-when
+class TypeHintTests(TestCase):
+    def test_returning_type_hint(self):
+        class MyClass:
+            def get_name(self) -> str:
+                return 'a name'
+
+        with Stub(MyClass) as my_class:
+            my_class.get_name().returns('another name')
+
+
 # https://bitbucket.org/DavidVilla/python-doublex/issues/24/customize-different-return-values-for
 class function_doubles(TestCase):
     def test_hack_function_spy(self):
