--- doublex/test3/test
@@ -1,1 +1,0 @@
-../test
